requirements.txt
ai71==0.0.21
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
cerebras_cloud_sdk==1.14.0
certifi==2024.7.4
click==8.1.7
distro==1.9.0
dnspython==2.7.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
httpx-sse==0.3.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
openai==1.35.10
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.10.1
PyPDF2==3.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
tqdm==4.66.4
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
python-dotenv
